ripped out rapi.search, no more search_backend intermediary
